backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
torch==2.3.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
joblib==1.4.2
imbalanced-learn==0.12.2
python-dotenv==1.0.1
structlog==24.1.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
aiofiles==23.2.1
pyarrow>=15.0.0
